% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R, R/exposure-package.R
\docType{package}
\name{exposure}
\alias{exposure}
\title{exposure}
\usage{
exposure(net_p, net_v, dispersal = 20)
}
\arguments{
\item{net_p}{A \pkg{dodgr} network weighted for pedestrian transport}

\item{net_v}{A \pkg{dodgr} network weighted for vehicular transport with an
additional \code{centrality} column appended by \link{exposure_centrality}.}

\item{dispersal}{Range in metres over which vehicular emissions are
effectively dispersed.}
}
\value{
A \code{dodgr} network with additional columns of 'centrality-veh' for
vehicular centrality; 'centrality-ped' for pedestrian centrality; and
'exposure' for relative risk from exposure of pedestrians to vehicular
emissions.
}
\description{
Calculate time-based betweenness centrality for a street network, as routed
either for vehicular or pedestrian transport.

Spatial relative risk from exposure to urban pollutants.
}
