% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{exposure_centrality}
\alias{exposure_centrality}
\title{exposure_centrality}
\usage{
exposure_centrality(graph, pedestrian = TRUE)
}
\arguments{
\item{graph}{'data.frame' or equivalent object representing the network
graph (see Details)}

\item{pedestrian}{If \code{TRUE}, calculate pedestrian-weighted centrality,
otherwise centrality for vehicular (motorcar) transport.}
}
\value{
Modified version of graph with additonal 'centrality' column added.
}
\description{
Calculate time-based betweenness centrality for a street network, as routed
either for vehicular or pedestrian transport.
}
