% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{exposure_disperse}
\alias{exposure_disperse}
\title{exposure_disperse}
\usage{
exposure_disperse(graph, from, dens, k = 500, tol = 0.000000000001)
}
\arguments{
\item{graph}{\code{data.frame} or equivalent object representing the network
graph, with one column quantifying pollutant concentrations throughout the
network.}

\item{from}{Vector or matrix of points \strong{from} which aggregate dispersed
flows are to be calculated (see Details)}

\item{dens}{Vectors of densities correponsing to the \code{from} points}

\item{k}{Width coefficient of exponential diffusion function defined as
\code{exp(-d/k)}, in units of distance column of \code{graph} (metres by default).}

\item{tol}{Relative tolerance below which dispersal is considered to have
finished. This parameter can generally be ignored; if in doubt, its effect
can be removed by setting \code{tol = 0}.}
}
\value{
Modified version of graph with additonal \code{exposure} column added.
}
\description{
Disperse flows throughout a network based on a input vectors of origin points
and associated densities, and calculate resultant relative risk of exposure
to urban pollutants.
}
