% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{exposure_centrality}
\alias{exposure_centrality}
\title{exposure_centrality}
\usage{
exposure_centrality(net, dispersal = 20)
}
\arguments{
\item{net}{An \code{sc}-class street network extracted with
\code{dodgr::dodgr_streetnet_sc}.}

\item{dispersal}{Range in metres over which vehicular emissions are
effectively dispersed.}
}
\value{
A \code{dodgr} network with additional columns of 'centrality-veh' for
vehicular centrality; 'centrality-ped' for pedestrian centrality; and
'exposure' for relative risk from exposure of pedestrians to vehicular
emissions.
}
\description{
Calculate time-based betweenness centrality for a street network, as routed
either for vehicular or pedestrian transport.
}
