% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{exposure_centrality}
\alias{exposure_centrality}
\title{exposure_centrality}
\usage{
exposure_centrality(net)
}
\arguments{
\item{net}{A \pkg{dodgr} network weighted for vehicular transport (that is,
with \code{wt_profile = "motorcar"}).}
}
\value{
Modified version of same net with additional "centrality" column
appended.
}
\description{
Calculate centrality for a street network
}
